pyproject.toml
[tool.poetry]
name = "handson-bing-spell-check"
version = "0.1.0"
description = ""
authors = ["Takuya Sato <business@satotakuya.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
